# docs/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Servicio centralizado de autenticación y autorización con JWT RS256, RBAC y auditoría.",
        "title": "Sentinel API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/auth/change-password": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "AppKeyAuth": []
                    }
                ],
                "description": "Cambia la contraseña del usuario autenticado. Requiere la contraseña actual.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cambiar contraseña",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Clave secreta de la aplicación",
                        "name": "X-App-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token JWT. Formato: Bearer {token}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Contraseñas actual y nueva",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Contraseña cambiada exitosamente"
                    },
                    "400": {
                        "description": "Datos inválidos o política de contraseña",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "security": [
                    {
                        "AppKeyAuth": []
                    }
                ],
                "description": "Autentica un usuario con credenciales y retorna tokens JWT de acceso y refresco.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Clave secreta de la aplicación",
                        "name": "X-App-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Credenciales de acceso",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Autenticación exitosa",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerLoginResponse"
                        }
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Credenciales incorrectas o aplicación no encontrada",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Cuenta inactiva o bloqueada",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "AppKeyAuth": []
                    }
                ],
                "description": "Invalida el token de refresco del usuario autenticado.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Clave secreta de la aplicación",
                        "name": "X-App-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token JWT. Formato: Bearer {token}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sesión cerrada exitosamente"
                    },
                    "401": {
                        "description": "No autenticado",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "security": [
                    {
                        "AppKeyAuth": []
                    }
                ],
                "description": "Renueva el token de acceso usando un token de refresco válido.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autenticación"
                ],
                "summary": "Renovar tokens",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Clave secreta de la aplicación",
                        "name": "X-App-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Token de refresco",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerRefreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tokens renovados exitosamente",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerTokenResponse"
                        }
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Token inválido, expirado o revocado",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_handler.SwaggerChangePasswordRequest": {
            "type": "object",
            "properties": {
                "current_password": {
                    "type": "string",
                    "example": "Admin@Local1!"
                },
                "new_password": {
                    "type": "string",
                    "example": "NuevaClave@2025!"
                }
            }
        },
...
        "internal_handler.SwaggerErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "details": {},
                "message": {
                    "type": "string",
                    "example": "campo requerido"
                }
            }
        },
        "internal_handler.SwaggerErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/internal_handler.SwaggerErrorDetail"
                }
            }
        },
        "internal_handler.SwaggerLoginRequest": {
            "type": "object",
            "properties": {
                "client_type": {
                    "type": "string",
                    "enum": [
                        "web",
                        "mobile",
                        "desktop"
                    ],
                    "example": "web"
                },
                "password": {
                    "type": "string",
                    "example": "Admin@Local1!"
                },
                "username": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
        "internal_handler.SwaggerLoginResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "expires_in": {
                    "type": "integer",
                    "example": 3600
                },
                "refresh_token": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "token_type": {
                    "type": "string",
                    "example": "Bearer"
                },
                "user": {
                    "$ref": "#/definitions/internal_handler.SwaggerLoginUser"
                }
            }
        },
        "internal_handler.SwaggerLoginUser": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "admin@empresa.com"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "must_change_password": {
                    "type": "boolean",
                    "example": false
                },
                "username": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
...
        "internal_handler.SwaggerRefreshRequest": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
        "internal_handler.SwaggerTokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "expires_in": {
                    "type": "integer",
                    "example": 3600
                },
                "refresh_token": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "token_type": {
                    "type": "string",
                    "example": "Bearer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AppKeyAuth": {
            "description": "Clave secreta de la aplicación cliente",
            "type": "apiKey",
            "name": "X-App-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Token JWT de acceso. Formato: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}